adressing code comments
